fix test and mixing as per suggestions
