+++ src/aap_eda/core/models/mixins.py
@@ -1,5 +1,18 @@
+#  Copyright 2023 Red Hat, Inc.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
 from django.core.exceptions import FieldDoesNotExist
-from django.utils import timezone
 
 
 class ModifiedAtUpdaterMixin:
@@ -21,5 +34,5 @@ class ModifiedAtUpdaterMixin:
             and field.auto_now
             and kwargs.get("update_fields", [])
         ):
-            self.modified_at = timezone.now()
+            kwargs["update_fields"].append("modified_at")
         super().save(*args, **kwargs)
